broke out Global styles to its own style component
